+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:openscan/Utilities/constants.dart';
 import 'package:openscan/presentation/screens/about_screen.dart';
-import 'package:openscan/presentation/screens/getting_started_screen.dart';
+import 'package:openscan/presentation/screens/demo_screen.dart';
 import 'package:openscan/presentation/screens/home_screen.dart';
 import 'package:openscan/presentation/screens/splash_screen.dart';
 import 'package:openscan/presentation/screens/view_document.dart';
