+++ lib/presentation/Widgets/view/quality_selector.dart
