+++ lib/presentation/screens/splash_screen.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:openscan/Utilities/constants.dart';
-import 'package:openscan/presentation/screens/getting_started_screen.dart';
+import 'package:openscan/presentation/screens/demo_screen.dart';
 import 'package:openscan/presentation/screens/home_screen.dart';
 import 'package:shared_preferences/shared_preferences.dart';
